# FIX-#2311: fixed performance regression at reduction operations