--- a/modin/backends/pandas/query_compiler.py
+++ b/modin/backends/pandas/query_compiler.py
@@ -171,8 +171,9 @@ def _numeric_only_reduce_fn(applier: Type[Function], *funcs) -> Callable:
     """
 
     def caller(self, *args, **kwargs):
-        # If `numeric_only` is None then we don't know what columns/indices will
-        # be dropped at the result of reduction function, and so can't preserve labels
+        # If `numeric_only` is None and the frame contains non-numeric columns,
+        # then we don't know what columns/indices will be dropped at the result
+        # of reduction function, and so can't preserve labels
         preserve_index = kwargs.get("numeric_only", None) is not None
         return applier.register(*funcs, preserve_index=preserve_index)(
             self, *args, **kwargs