@YarShev 2020-10-27
Can you measure two more functions: `sum_min_count` and `prod_min_count` and add measurements for all the functions for axis=1? Also, please attention for comments in [the commit](https://github.com/modin-project/modin/commit/55e345951818761ad26711cde91bb0d2b296aa5f).