@dchigarev 2020-10-27
@YarShev added new time measurements at the [header of this PR](https://github.com/modin-project/modin/pull/2314#issue-510056226) and answered all comments for [this commit](https://github.com/modin-project/modin/commit/55e345951818761ad26711cde91bb0d2b296aa5f)